Move AbstractExecutor into green module
