+++ tango/tango_futures.py
@@ -16,7 +16,7 @@ from __future__ import absolute_import
 from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
 
 # Tango imports
-from .tango_executor import AbstractExecutor
+from .green import AbstractExecutor
 
 __all__ = ["FuturesExecutor", "get_global_executor", "set_global_executor"]
